chore: addresing PR feedback
