+++ app/default_overrides.go
@@ -154,14 +154,14 @@ func (icaModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
 	return cdc.MustMarshalJSON(gs)
 }
 
-// pfmMiddleware is a wrapper around packetforward.AppModule which adds required no-op migrations for upgrade.
-type pfmMiddleware struct {
+// pfm is a wrapper around packetforward.AppModule which adds required no-op migrations for upgrade.
+type pfm struct {
 	packetforward.AppModule
 }
 
 // RegisterServices needs to be overridden to add a no-op handler when going from v1 -> v2
 // the existing app module (v8) has a built-in migration for going from v2 -> v3
-func (am pfmMiddleware) RegisterServices(cfg module.Configurator) {
+func (am pfm) RegisterServices(cfg module.Configurator) {
 	err := cfg.RegisterMigration(packetforwardtypes.ModuleName, 1, func(sdk.Context) error {
 		// a no-op registration needs to happen from v1 -> v2.
 		return nil
